refactor to redefine result module
